cor(kuri) : les valeurs des énums z64 et n64 sont tronquées
Les valeurs étaient stockées dans un entier sur 32 bits, donc les
valeurs sur 64 bits furent tronquées.

+++ logiciels/kuri/arbre_syntaxique/utilitaires.hh
@@ -447,7 +447,7 @@ struct MembreTypeComposé {
     Type *type = nullptr;
     IdentifiantCode *nom = nullptr;
     unsigned decalage = 0;
-    int valeur = 0;                                       // pour les énumérations
+    uint64_t valeur = 0;                                  // pour les énumérations
     NoeudExpression *expression_valeur_defaut = nullptr;  // pour les membres des structures
     int drapeaux = 0;
     uint32_t rembourrage = 0;
